debug

+++ nafion layer electrical field/SolverIndex.cpp
@@ -4,23 +4,23 @@ OneDIndex::OneDIndex(const mesh& membrane, const mesh& solution)
 {
 	InnerIndex.resize(SpeciesEnum::Count);
 
-	for (int k = SpeciesEnum::mReactant; k < SpeciesEnum::sReactant; ++k) {
+	for (long k = SpeciesEnum::mReactant; k < SpeciesEnum::sReactant; ++k) {
 		InnerIndex[k].resize(membrane.m);
-		for (int i = 0; i < membrane.m; ++i) {
+		for (long i = 0; i < membrane.m; ++i) {
 			InnerIndex[k][i].resize(membrane.n);
 
-			for (int j = 0; j < membrane.n; ++j) {
+			for (long j = 0; j < membrane.n; ++j) {
 				InnerIndex[k][i][j] = i*membrane.n + j + k*membrane.Getmxn();
 			}
 		}
 	}
 
-	for (int k = SpeciesEnum::sReactant; k < SpeciesEnum::Count; ++k) {
+	for (long k = SpeciesEnum::sReactant; k < SpeciesEnum::Count; ++k) {
 		InnerIndex[k].resize(solution.m);
-		for (int i = 0; i < solution.m; ++i) {
+		for (long i = 0; i < solution.m; ++i) {
 			InnerIndex[k][i].resize(solution.n);
 
-			for (int j = 0; j < solution.n; ++j) {
+			for (long j = 0; j < solution.n; ++j) {
 				InnerIndex[k][i][j] = i*membrane.n + j + membrane.Getmxn()*SpeciesEnum::sReactant + solution.Getmxn()*(k - SpeciesEnum::sReactant);
 			}
 		}
@@ -29,24 +29,24 @@ OneDIndex::OneDIndex(const mesh& membrane, const mesh& solution)
 
 TwoDIndex::TwoDIndex(const mesh& membrane, const mesh& solution)
 {
-	unsigned long lenth = membrane.Getmxn()*SpeciesEnum::sReactant + solution.Getmxn()*(SpeciesEnum::Count - SpeciesEnum::sReactant);
+	long lenth = membrane.Getmxn()*SpeciesEnum::sReactant + solution.Getmxn()*(SpeciesEnum::Count - SpeciesEnum::sReactant);
 	InnerIndex.resize(lenth);
 
-	for (int k = SpeciesEnum::mReactant; k < SpeciesEnum::sReactant; ++k) {
-		for (int i = 0; i < membrane.m; ++i) {
-			for (int j = 0; j < membrane.n; ++j) {
+	for (long k = SpeciesEnum::mReactant; k < SpeciesEnum::sReactant; ++k) {
+		for (long i = 0; i < membrane.m; ++i) {
+			for (long j = 0; j < membrane.n; ++j) {
 
-				unsigned long index = i*membrane.n + j + k*membrane.Getmxn();
+				long index = i*membrane.n + j + k*membrane.Getmxn();
 				InnerIndex[index] = std::make_tuple(k, j, i);
 			}
 		}
 	}
 
-	for (int k = SpeciesEnum::sReactant; k < SpeciesEnum::Count; ++k) {
-		for (int i = 0; i < solution.m; ++i) {
-			for (int j = 0; j < solution.n; ++j) {
+	for (long k = SpeciesEnum::sReactant; k < SpeciesEnum::Count; ++k) {
+		for (long i = 0; i < solution.m; ++i) {
+			for (long j = 0; j < solution.n; ++j) {
 
-				unsigned long index = i*membrane.n + j + membrane.Getmxn()*SpeciesEnum::sReactant + solution.Getmxn()*(k - SpeciesEnum::sReactant);
+				long index = i*membrane.n + j + membrane.Getmxn()*SpeciesEnum::sReactant + solution.Getmxn()*(k - SpeciesEnum::sReactant);
 				InnerIndex[index] = std::make_tuple(k, j, i);
 			}
 		}
